+++ pyee/__init__.py
@@ -33,6 +33,7 @@ except ImportError:
     ensure_future = None
 
 from collections import defaultdict
+from functools import wraps
 
 __all__ = ['EventEmitter', 'PyeeException']
 
@@ -104,6 +105,10 @@ class EventEmitter(object):
         This will automatically schedule the coroutine using the configured
         scheduling function (defaults to ``asyncio.ensure_future``) and the
         configured event loop (defaults to ``asyncio.get_event_loop()``).
+
+        In both the decorated and undecorated forms, the function handler
+        is returned. The upshot of this is that you can call decorated handlers
+        directly, as well as use them in remove_listener calls.
         """
 
         def _on(f):
@@ -113,7 +118,6 @@ class EventEmitter(object):
             # Add the necessary function
             self._events[event].append(f)
 
-            # Return original function so removal works
             return f
 
         if f is None:
@@ -176,8 +180,35 @@ class EventEmitter(object):
     def once(self, event, f=None):
         """The same as ``ee.on``, except that the listener is automatically
         removed after being called.
+
+        A subtle difference in behavior between the decorated and
+        undecorated versions of this function: As an implementation detail,
+        the event handler is wrapped in a function which removes itself from
+        listeners right before the wrapped listener is called. The decorator
+        version returns the underlying function, while the non-decorating call
+        returns the wrapped handler. This means that this will work::
+
+            h = ee.once('event', once_handler)
+
+            ee.remove_listener('event', h)
+
+        However::
+
+            @ee.once('event')
+            def once_handler():
+                print('hello')
+
+            # This won't remove anything, since once_handler isn't the
+            # wrapped listener
+            ee.remove_listener('event', once_handler)
+
+            # But, I can call it without removing the event
+            once_handler()  # Prints "hello", event still attached
+            ee.emit('event')  # Still prints "hello"
+            ee.emit('event')  # Does nothing, event self-removed
         """
         def _once(f):
+            @wraps(f)
             def g(*args, **kwargs):
                 self.remove_listener(event, g)
                 # f may return a coroutine, so we need to return that
@@ -185,14 +216,14 @@ class EventEmitter(object):
                 return f(*args, **kwargs)
             return g
 
-        def _wrapper(f):
+        def _decorator(f):
             self.on(event, _once(f))
             return f
 
         if f is None:
-            return _wrapper
+            return _decorator
         else:
-            _wrapper(f)
+            return self.on(event, _once(f))
 
     def remove_listener(self, event, f):
         """Removes the function ``f`` from ``event``."""